Implemented Save for todo
Functionalities to be implemented are: show active todos, styling of app
