--- app/component.js
@@ -1,7 +1,0 @@
-module.exports = () => {
-  const element= document.createElement('h1');
-
-  element.innerHTML = 'Kusan world';
-
-  return element;
-};
